+++ src/main/java/com/wakeUpTogetUp/togetUp/group/GroupService.java
@@ -1,7 +1,7 @@
 package com.wakeUpTogetUp.togetUp.group;
 
-import com.wakeUpTogetUp.togetUp.common.BaseException;
-import com.wakeUpTogetUp.togetUp.common.BaseResponseStatus;
+import com.wakeUpTogetUp.togetUp.common.ResponseStatus;
+import com.wakeUpTogetUp.togetUp.common.exception.BaseException;
 import com.wakeUpTogetUp.togetUp.group.dto.request.GroupReq;
 import com.wakeUpTogetUp.togetUp.group.dto.response.GroupRes;
 import com.wakeUpTogetUp.togetUp.group.model.Group;
@@ -52,7 +52,7 @@ public class GroupService {
     // 그룹수정 이름, 인트로
         Optional<Group> group = Optional.ofNullable(groupRepository.findById(groupId)
                 .orElseThrow(
-                        () -> new BaseException(BaseResponseStatus.INVALID_GROUP_ID)
+                        () -> new BaseException(ResponseStatus.INVALID_GROUP_ID)
                 ));
         group.get().setName(groupReq.getName());
         group.get().setIntro(groupReq.getIntro());
